% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeA.R
\name{makeA}
\alias{makeA}
\title{Create Molecule Metadata}
\usage{
makeA(Go = TRUE)
}
\arguments{
\item{Go}{Set to TRUE to create the list}
}
\value{
\item{A}{A list containing the required meta data to complete a protein molecule}
}
\description{
\code{makeA} Produces a list containing all of the meta data necessary
to complete a protein molecule.
}
\details{
Creates a list containing the following information about the amino acid structure: \cr
anchors \cr
atom \cr
cliqs \cr
cliqs_woh \cr
dihedral \cr
first_meta \cr
info \cr
last_meta \cr
meta_bonds \cr
omission_map \cr
X \cr

All of the raw data is available in the package source
}
\examples{
A <- makeA()

}

