% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{mst}
\alias{mst}
\title{Compute Minimum Spanning Tree}
\usage{
mst(D)
}
\arguments{
\item{D}{A distance matrix}
}
\value{
MST a data frame object of 3 columns containing the parent nodes, child nodes, and corresponding weight of the MST edge
}
\description{
\code{mst} Compute a minimum spanning tree using Prim's algorithm
}
\examples{

X <- runif(10,0,1)
Y <- runif(10,0,1)
D <- dist(cbind(X,Y))

mst(as.matrix(D))

}
